+++ demos/react-demo/src/index.tsx
@@ -18,6 +18,18 @@ import reportWebVitals from './reportWebVitals';
 import './index.css';
 import '@brightlayer-ui/icons/BrightlayerUIIcons.css';
 
+// TODO: Remove this after the issues with @types/react goes away
+// https://github.com/DefinitelyTyped/DefinitelyTyped/discussions/68444
+/* eslint-disable */
+declare global {
+    namespace React {
+        interface DOMAttributes<T> {
+            placeholder?: string | undefined;
+        }
+    }
+}
+/* eslint-enable */
+
 const container = document.getElementById('root');
 if (!container) throw new Error('Root Element was not found in the DOM');